requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
autopep8==1.7.0
backports.zoneinfo==0.2.1
cffi==1.15.1
cryptography==37.0.4
Django==4.1
django-environ==0.9.0
djangorestframework==3.12.1
djangorestframework-simplejwt==5.2.0
mysqlclient==2.1.1
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.4.0
pytz==2022.2.1
sqlparse==0.4.2
toml==0.10.2
